requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord.py==2.4.0
Flask==3.0.3
frozenlist==1.4.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lyricsgenius==3.0.1
MarkupSafe==2.1.5
multidict==6.0.5
mutagen==1.47.0
prettytable==3.10.2
pycparser==2.22
pycryptodomex==3.20.0
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.5
urllib3==2.2.2
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
yarl==1.9.4
yt-dlp==2024.8.1
